+++ CHANGES.md
@@ -37,6 +37,7 @@
 - Use `pathlib.Path` instead of `os.path`.
 
 - Support `--update` flag in `mxmake init` command, updating the Makefile without prompting for settings.
+- Use importlib.metadata to load entrypoints.
 
 ## 1.0a3 (2024-02-06)